+++ src/Repositories/Interfaces/CoreRepositoryInterface.php
@@ -3,10 +3,12 @@
 namespace PacerIT\LaravelCore\Repositories\Interfaces;
 
 use Illuminate\Contracts\Container\BindingResolutionException;
+use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\Collection;
 use PacerIT\LaravelCore\Entities\Interfaces\CoreEntityInterface;
 use PacerIT\LaravelCore\Repositories\Criteria\Interfaces\CoreRepositoryCriteriaInterface;
 use PacerIT\LaravelCore\Repositories\Exceptions\RepositoryEntityException;
+use Yajra\DataTables\EloquentDataTable;
 
 /**
  * Interface CoreRepositoryInterface
@@ -41,11 +43,11 @@ interface CoreRepositoryInterface
     /**
      * Get entity instance
      *
-     * @return CoreEntityInterface
+     * @return CoreEntityInterface|Builder
      * @author Wiktor Pacer <kontakt@pacerit.pl>
      * @since 2019-07-05
      */
-    public function getEntity(): CoreEntityInterface;
+    public function getEntity();
 
     /**
      * Push criteria
@@ -175,5 +177,26 @@ interface CoreRepositoryInterface
      * @since 2019-07-10
      */
     public function firstOrNull(array $where): ?CoreEntityInterface;
+
+    /**
+     * Make datatable response
+     *
+     * @param array $columns
+     * @return EloquentDataTable
+     * @author Wiktor Pacer <kontakt@pacerit.pl>
+     * @since 2019-07-31
+     */
+    public function datatable(array $columns): EloquentDataTable;
+
+    /**
+     * Order by records
+     *
+     * @param string $column
+     * @param string $direction
+     * @return mixed
+     * @author Wiktor Pacer <kontakt@pacerit.pl>
+     * @since 2019-07-31
+     */
+    public function orderBy(string $column, string $direction = 'asc');
 
 }