+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # Changelog
+## v.1.0.2
+    - update documentation
+    - extend CoreRepository
+        * add datatable() function, producing datatable response
+        * add orderBy() function
+        * fix getEntity() function
+    - add Criteria
+        * DateCriteria
+        * FindWhereCriteria
+        * FindWhereInCriteria
+        * LimitCriteria
+        * OrWhereCriteria
+        * Select2Criteria
 ## v.1.0.1
     - update documentation
     - add NumberTrait class